# docs/swagger.yaml
basePath: /
definitions:
  handlers.BlockRequest:
    properties:
      userId:
        example: 1
        type: integer
    type: object
  handlers.FollowRequest:
    properties:
      userId:
        example: 1
        type: integer
    type: object
  handlers.LoginRequestBody:
    properties:
      password:
        example: P@ssw0rd
        type: string
      username:
        example: username
        type: string
    type: object
  handlers.MessageResponseBody:
    properties:
      content:
        example: bla-bla-bla
        type: string
      recipientId:
        example: 1
        type: integer
    type: object
  handlers.RefreshTokenRequestBody:
    properties:
      refreshToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
    type: object
  handlers.RegistrationRequestBody:
    properties:
      password:
        example: P@ssw0rd
        type: string
      username:
        example: username
        type: string
    type: object
  handlers.Response:
    properties:
      data: {}
      error:
        example: false
        type: boolean
      message:
        example: success operation
        type: string
    type: object
  handlers.TokensResponseBody:
    properties:
      accessToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
      refreshToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
    type: object
  handlers.UpdatePasswordRequestBody:
    properties:
      newPassword:
        example: NewP@ssw0rd
        type: string
      oldPassword:
        example: OldP@ssw0rd
        type: string
    type: object
  handlers.UpdateUsernameRequestBody:
    properties:
      username:
        example: username
        type: string
    type: object
  models.User:
    properties:
      blockedUsers:
        items:
          $ref: '#/definitions/models.User'
        type: array
      followedUsers:
        items:
          $ref: '#/definitions/models.User'
        type: array
      id:
        example: 1
        type: integer
      username:
        example: username
        type: string
    type: object
host: localhost:80
info:
  contact:
    name: Artem Kostenko
    url: https://github.com/aerosystems
  description: Backend App for simple social Live Chat
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 'NIX Junior: Chat App'
  version: "1.0"
paths:
  /v1/search:
    get:
      consumes:
      - application/json
      description: Search users by username with autocomplete
      parameters:
      - description: query string for search by username, minimum 3 characters, maximum
          40 characters
        in: query
        name: q
        required: true
        type: string
      - description: 'type of search, default: ''user'', available: ''user'', ''friend'',
          ''blacklist'''
        in: query
        name: type
        type: string
      - description: 'order of search, default: ''asc'', available: ''asc'', ''desc'''
        in: query
        name: order
        type: string
      - description: 'limit of search, default: ''10'', available: ''1-1000'''
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: search users
      tags:
      - search
  /v1/token/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: raw request body, should contain Refresh Token
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequestBody'
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.TokensResponseBody'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: refresh pair JWT tokens
      tags:
      - auth
  /v1/user:
    get:
      consumes:
      - application/json
      description: Get user data
      parameters:
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: Get user data
      tags:
      - user
  /v1/user/block:
    post:
      consumes:
      - application/json
      parameters:
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: raw request body
        in: body
        name: follow
        required: true
        schema:
          $ref: '#/definitions/handlers.BlockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: Block user
      tags:
      - relationship
  /v1/user/follow:
    post:
      consumes:
      - application/json
      parameters:
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: raw request body
        in: body
        name: follow
        required: true
        schema:
          $ref: '#/definitions/handlers.FollowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: Follow user
      tags:
      - relationship
  /v1/user/login:
    post:
      consumes:
      - application/json
      description: |-
        Username should contain:
        - lower, upper case latin letters and digits
        - minimum 8 characters length
        - maximum 40 characters length
        Password should contain:
        - minimum of one small case letter
        - minimum of one upper case letter
        - minimum of one digit
        - minimum of one special character
        - minimum 8 characters length
        - maximum 40 characters length
        Response contain pair JWT tokens, use /v1/tokens/refresh for updating them
      parameters:
      - description: raw request body
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.TokensResponseBody'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: login user by credentials
      tags:
      - auth
  /v1/user/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: logout user
      tags:
      - auth
  /v1/user/register:
    post:
      consumes:
      - application/json
      description: |-
        Username should contain:
        - lower, upper case latin letters and digits
        - minimum 8 characters length
        - maximum 40 characters length
        Password should contain:
        - minimum of one small case letter
        - minimum of one upper case letter
        - minimum of one digit
        - minimum of one special character
        - minimum 8 characters length
        - maximum 40 characters length
      parameters:
      - description: raw request body
        in: body
        name: registration
        required: true
        schema:
          $ref: '#/definitions/handlers.RegistrationRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: registration user by credentials
      tags:
      - auth
  /v1/user/update-password:
    put:
      consumes:
      - application/json
      description: |-
        OldPassword/NewPassword should contain:
        - minimum of one small case letter
        - minimum of one upper case letter
        - minimum of one digit
        - minimum of one special character
        - minimum 8 characters length
        - maximum 40 characters length
      parameters:
      - description: raw request body
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdatePasswordRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: update password
      tags:
      - user
  /v1/user/update-username:
    put:
      consumes:
      - application/json
      description: |-
        Username should contain:
        - lower, upper case latin letters and digits
        - minimum 8 characters length
        - maximum 40 characters length
      parameters:
      - description: raw request body
        in: body
        name: username
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUsernameRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: update username
      tags:
      - user
  /ws/chat:
    get:
      description: Chat with users based on WebSocket
      parameters:
      - description: should contain Access Token, with the Bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: body should contain content and recipient_id for sending message
        in: body
        name: chat
        required: true
        schema:
          $ref: '#/definitions/handlers.MessageResponseBody'
      responses:
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Response'
      summary: Chat [WebSocket]
      tags:
      - chat
swagger: "2.0"